+++ CHANGELOG.md
@@ -63,6 +63,8 @@
     - cancel command substitution by SIGINT
 - **Breaking Change**: during sub-shell creation, not clear job entry
     - for builtin ``jobs`` command within sub-shell
+- **Breaking Change**: after background job termination, not show signal message
+    - not show signal messages after call ``Job#poll``, ``Job#wait``, builtin wait command
 - show job information via CTRL-Z/``fg``/``bg``
 - escape unprintable character when throw some errors